Main timer fixed
In order to reduce the discrepancy between the displayed time elapsed and actual time elapsed (as the displayed time elapsed ordinarily would count approximately 2 seconds ahead for each actual minute that has elapsed) the timing of the first second increment of every minute increment (not when the timer initially starts counting) is tripled in length. So far, this change has only been applied to the main timer.

In addition, it will be important to test the code on the device of which it is intended for use. Both the overall visual format and timers' accuracies will likely be affected by the chosen device's display dimensions and general performance. Other potential tweaks related to the effectiveness of the program as a whole may become apparent through such efforts as well.